feat: Implement persistent storage with docker volumes and default layout loading and layout_svg rendering
