+++ Dockerfile
@@ -2,8 +2,8 @@ FROM python:3.11-slim
 
 WORKDIR /server
 
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
+COPY server_requirements.txt .
+RUN pip install --no-cache-dir -r server_requirements.txt
 
 COPY server/ .
